+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "Bot.js",
   "scripts": {
-    "test": "mocha ./src/Test/test.js",
+    "test": "mocha ./src/test",
     "dev": "nodemon ./src/Bot.js",
     "web": "node ./src/Bot.js",
     "start": "npm run dev"
